Make LineNuber and SourcePath top level classes
