+++ compiler/src/dotty/tools/dotc/core/Definitions.scala
@@ -1084,6 +1084,7 @@ class Definitions {
 
   // ----- Sources ----------------------------------------------------------
 
-  lazy val DottySourcePositionModuleRef = ctx.requiredModuleRef("dotty.source.Position")
+  lazy val DottySourceLineNumberModuleClass = ctx.requiredModuleRef("dotty.source.position.LineNumber").termSymbol.moduleClass
+  lazy val DottySourceSourcePathModuleClass = ctx.requiredModuleRef("dotty.source.position.SourcePath").termSymbol.moduleClass
 
 }
